Refactor JS into modules, add Jest tests, fix repository health bugs
Split the monolithic generate-dashboard.js (927 lines) into focused modules:
- src/config.js: configuration loading
- src/utils.js: pure utility functions with testable interfaces
- src/github-api.js: GitHub API fetchers with fallback query support
- src/data-processing.js: stats computation, priority scoring, data collection
- src/html-generators.js: HTML section generators

Bug fixes:
- Add fallback GraphQL query without vulnerabilityAlerts for orgs where
  the token lacks security_events permission (fixes missing mage-os org data)
- Repository Health and Missing Mirrors now iterate ALL configured orgs
  instead of being hardcoded to only 'mage-os' (fixes empty sections)
- Consolidate 3 inconsistent date formatting patterns into formatDateUTC()

Testing:
- Add Jest with ESM support (66 tests across 3 test files)
- Tests cover utilities, data processing logic, and HTML badge helpers
- Add GitHub Actions workflow to run tests on PRs
