+++ jest.config.js
@@ -1,0 +1,4 @@
+export default {
+    transform: {},
+    testMatch: ['**/tests/**/*.test.js'],
+};
