+++ package.json
@@ -5,7 +5,8 @@
   "private": true,
   "scripts": {
     "build": "node generate-dashboard.js",
-    "report": "node generate-report.js"
+    "report": "node generate-report.js",
+    "test": "node --experimental-vm-modules node_modules/.bin/jest"
   },
   "engines": {
     "node": ">=18.0.0"
@@ -13,5 +14,8 @@
   "dependencies": {
     "escape-html": "^1.0.3",
     "tinycolor2": "^1.6.0"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0"
   }
 }
